Move loading functions to Loader class
